substrate/requirements.txt
typedb-driver>=3.0.0,<4
